// src/components/Filters/interfaces.tsx
export interface CurrencyOption {
    id: string;
    name: string;
    checked: boolean;
}
export interface MultiplierOptions {
    id: string;
    name: string;
    checked: boolean;
}
export interface PriceOptions {
    id: string;
    name: string;
    checked: boolean;
}
export interface MultiplierTypeOptions {
    id: string;
    name: string;
    checked: boolean;
}

export interface FilterProps {
    name?: string;
}